% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knapsack_cpp}
\alias{knapsack_cpp}
\title{C++ implementation of the knapsack problem}
\usage{
knapsack_cpp(x, W)
}
\arguments{
\item{x}{A data frame with two variables (w, v) that contains the weight (w) and value (v) of each item.}

\item{W}{The knapsack capacity.}
}
\description{
C++ implementation of the knapsack problem
}
